Change coordinates data type from double to float

+++ app/src/main/java/com/github/dtcan/weatherapp/api/City.java
@@ -6,9 +6,9 @@ import org.json.JSONObject;
 public class City {
 
     public final String name;
-    public final double lat, lon;
+    public final float lat, lon;
 
-    public City(String name, double lat, double lon) {
+    public City(String name, float lat, float lon) {
         this.name = name;
         this.lat = lat;
         this.lon = lon;
@@ -18,8 +18,8 @@ public class City {
         String name = cityJSON.getString("name");
         String state = cityJSON.optString("state");
         String country = cityJSON.getString("country");
-        double lat = cityJSON.getDouble("lat");
-        double lon = cityJSON.getDouble("lon");
+        float lat = (float) cityJSON.getDouble("lat");
+        float lon = (float) cityJSON.getDouble("lon");
 
         if(state.length() > 0) {
             name += ", " + state;
